requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
aiohttp==3.1.1
aiopg==0.13.2
apistar[asyncio]==0.3.9
async-timeout==2.0.1      # via aiohttp
asyncio-redis==0.14.3     # via uvitools
attrs==17.4.0             # via aiohttp, pytest
certifi==2018.1.18        # via requests
chardet==3.0.4            # via aiohttp, requests
coreapi==2.3.3            # via apistar
coreschema==0.0.4         # via coreapi
gunicorn==19.7.1          # via uvicorn
httptools==0.0.10         # via uvicorn
idna-ssl==1.0.1           # via aiohttp
idna==2.6                 # via idna-ssl, requests, yarl
itypes==1.1.0             # via coreapi
jinja2==2.10              # via apistar, coreschema
markupsafe==1.0           # via jinja2
multidict==4.1.0          # via aiohttp, yarl
pluggy==0.6.0             # via pytest
psycopg2==2.7.4           # via aiopg
py==1.5.2                 # via pytest
pycryptodome==3.5.1
pyjwt==1.6.1
pytest==3.4.2             # via apistar
requests==2.18.4
ruamel.yaml==0.15.35
six==1.11.0               # via pytest
sqlalchemy==1.2.5
uritemplate==3.0.0        # via coreapi
urllib3==1.22             # via requests
uvicorn==0.0.15           # via apistar, uvitools
uvitools==0.0.4           # via apistar
uvloop==0.9.1             # via uvicorn
websockets==4.0.1         # via uvicorn
werkzeug==0.14.1          # via apistar, uvitools
whitenoise==3.3.1         # via apistar
yarl==1.1.1               # via aiohttp
